refactor: generalize const_expr reader
